hide edit/update buttons for non-owner users

+++ base/views.py
@@ -1,5 +1,8 @@
 from django.shortcuts import render , redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -8,6 +11,9 @@ from .form import RoomForm
 
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if  request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -53,6 +59,7 @@ def room(request,pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -63,10 +70,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request,'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
-    
+
+    if request.user != room.host:
+        return HttpResponse('you are not allowed here!')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -76,8 +87,12 @@ def updateRoom(request,pk):
     context ={'form': form}
     return render(request,'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('you are not allowed here!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
